LangGraph_projects/Chatbot/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "python-dotenv>=1.2.2",
]
